fix #6 (plugin fails to load)
This plugin failed to load under Asciidoctor 2.0.22 due to a difference
between the signature of the initialize method on the base
Asciidoctor::Extensions::TreeProcessor class and the implemented
initialize method in lib/asciidoctor/interdoc_reftext/processor.rb. This
commit modifies the function signature, and makes it so the Processor
class is now directly instantiated instead of by a reference passed to
the initialize method.